Add files via upload
This has the list view connected to the list of items in the resource file "elementsInArray.xml". When you click on the items in the list it will use toast to output the name of the item. I also have the "Make my Decision" Button placed properly. I am currently trying to get that button to output the randomly selected element using toast. I still need to find a way to allow the user to edit the items that are being displayed. After i get all of these core things done, i can add a second activity as the main screen that will have a button to take us to this screen.