Add files via upload
I was able to get the main menu working. It has a button that sends you to the list page. On the list page there are now three buttons. the green "make my decision" button that randomly selects and outputs an item using toast. A back button that sends us back to the home page. and a button with the text right that can be used for the edit function.